# Refactored header name matching for WASM target