Advantages:

 1. Shorter code
 2. More rusty and less macroy
 3. It should work in wasm now untested and make yew users happier
    https://github.com/DenisKolodin/yew/issues/478

(Un?)fortunately I also ran rustfmt on it - which changed the layouts for the tables a bit but not hideously.