@carllerche 2019-07-01
Hi, thanks for the work.

Would it be possible to include benchmarks for parsing header names? Probably could add a bench parsing each one + some common unknown header names. It would be interesting to compare the before and after.