@matiu2 2019-07-03
I tried to make it work in nightly and beta, but then it broke for rust 1.20.

The existing benches are pretty much the same, for example:

New: `test basic::insert_500_custom_headers::header_map            ... bench:      35,208 ns/iter (+/- 3,372)`

Old: `test basic::insert_500_custom_headers::header_map            ... bench:      36,283 ns/iter (+/- 573)`

I'll some benches just for the name parsing..