@matiu2 2019-07-05
I don't know how to bench unless it's running rust nightly (cargo bench) - so I had to make it work in nightly to get the numbers. It seems pretty close:

`cargo bench header_name`

# old macro style
test header_name_bad     ... bench:          15 ns/iter (+/- 0)
test header_name_easy    ... bench:          16 ns/iter (+/- 2)
test header_name_various ... bench:      10,222 ns/iter (+/- 991)

# new giant match statement
test header_name_bad     ... bench:          20 ns/iter (+/- 1)
test header_name_easy    ... bench:          21 ns/iter (+/- 1)
test header_name_various ... bench:       9,909 ns/iter (+/- 1,133)