@matiu2 2019-07-08
I've moved the make_all_known_headers out of the iterator. Sorry for not noticing that.

This commit is definitely slowing the code down by about 6% on my laptop :(

# New bench mark results

## Old code:

test header_name_bad     ... bench:          15 ns/iter (+/- 11)
test header_name_easy    ... bench:          16 ns/iter (+/- 1)
test header_name_various ... bench:       3,689 ns/iter (+/- 290)

## New code:

test header_name_bad     ... bench:          20 ns/iter (+/- 0)
test header_name_easy    ... bench:          24 ns/iter (+/- 1)
test header_name_various ... bench:       3,921 ns/iter (+/- 395)

--------

The main reason for this commit is so we can use `http` in wasm without an optimized build.

When debug building, those macros that expand out to `...&& b[0 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1] == b'p'...` get put as-is into wasm, causing the wasm implementiation to blow up at run-time trying to add all the ones.

Here's the issue in `yew` that I'm trying to fix: https://github.com/DenisKolodin/yew/issues/478#issuecomment-507429151