@matiu2 2019-07-09
Still slow. I don't know why :(

# Latest bench mark results

## Old macro based code:

test header_name_bad     ... bench:          15 ns/iter (+/- 0)
test header_name_easy    ... bench:          16 ns/iter (+/- 0)
test header_name_various ... bench:       3,648 ns/iter (+/- 123)

## New giant match statement code:

test header_name_bad     ... bench:          25 ns/iter (+/- 0)
test header_name_easy    ... bench:          24 ns/iter (+/- 0)
test header_name_various ... bench:       4,568 ns/iter (+/- 153)