@matiu2 2019-07-09
Telling it the len of the array makes it a bit faster:

test header_name_bad     ... bench:          25 ns/iter (+/- 0)
test header_name_easy    ... bench:          24 ns/iter (+/- 0)
test header_name_various ... bench:       4,566 ns/iter (+/- 206)