@matiu2 2019-07-09
# Latest bench mark results

## Old macro based code:

    $ cargo bench header_name
    ...
    test header_name_bad     ... bench:          14 ns/iter (+/- 0)
    test header_name_easy    ... bench:          16 ns/iter (+/- 1)
    test header_name_various ... bench:       3,763 ns/iter (+/- 80)

## New giant match statement code:

    test header_name_bad     ... bench:          25 ns/iter (+/- 1)
    test header_name_easy    ... bench:          24 ns/iter (+/- 0)
    test header_name_various ... bench:       4,227 ns/iter (+/- 75)