@carllerche 2019-07-10
We should definitely get this working with wasm. Perhaps there is a way to conditional compile a different fn in wasm?