@matiu2 2019-07-12
This version uses conditional compilation. It'll only compile the wasm version if there are 'debug_assertions' and the target is 'wasm32'.

It make the code longer, and harder to maintain though, because now if you add a new header, you'll have to add it in two places.

I think simplifying the code to the giant match statement is best. Although 8 more nanoseconds per header check may seem a lot; in msecs it is 8e-6 (0.0000008 msecs). But I'll leave the decision in your capable hands.

I tested it and it works in my wasm app in debug mode :)